+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Dependencies - Updated Credfeto.Version.Information.Generator to 1.0.43.281
 - Dependencies - Updated Credfeto.Date to 1.1.62.778
 - Dependencies - Updated Credfeto.Extensions.Linq to 1.0.66.649
+- Set cache headers for all expected responses
 ### Removed
 ### Deployment Changes